+++ include/api/include.php
@@ -1,6 +1,9 @@
 <?php
+    if(WeatherEnabled == true) {
+        require("include/api/OpenWeatherMap.php");
+    }
 
-require("include/api/OpenWeatherMap.php");
-require("include/api/newsapi.php");
-
+    if(NewsEnabled == true) { 
+        require("include/api/newsapi.php");
+    }
 ?>