+++ include/style/css.php
@@ -2,7 +2,7 @@
   <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" rel="stylesheet">
 
   <!-- Font Awesome -->
-  <link href="https://use.fontawesome.com/releases/v5.0.13/css/all.css" rel="stylesheet">
+  <link href="https://use.fontawesome.com/releases/v5.7.1/css/all.css" rel="stylesheet">
 
   <!-- Custom Styling -->
   <link href="<?php print(URL);?>/assets/css/custom.css" rel="stylesheet">