+++ links/mod.php
@@ -31,8 +31,12 @@
             $db_edit_linkcustomvar = stripslashes($home->GetLinkInfo($id)[0]['LinkCustomVar']);
             $db_edit_linkfontawesome = $home->GetLinkInfo($id)[0]['LinkFontAwesome'];
         } elseif($action =="edit" && $type == "header") {
-            header("location: " . URL . "/links/");
-            die();
+            $db_edit_hid = $home->GetHeaderInfo($id)[0]['HID'];
+            $db_edit_hname = $home->GetHeaderInfo($id)[0]['HeaderName'];
+            $db_edit_hlink = $home->GetHeaderInfo($id)[0]['HeaderLink'];
+            $db_edit_hddname = $home->GetHeaderInfo($id)[0]['HeaderDDName'];
+            $db_edit_horder = $home->GetHeaderInfo($id)[0]['HeaderOrder'];
+            $db_edit_hfonta = $home->GetHeaderInfo($id)[0]['HeaderFontAwesome'];
         } else {
             header("location: " . URL . "/links/");
             die();
@@ -47,7 +51,20 @@
             $db_new_linkcustomvar = addslashes($_POST['LinkCustomVar']);
             $db_new_linkfontawesome = htmlentities($_POST['LinkFontAwesome']);
             $home->UpdateLink($id, $db_new_link_hid, $db_new_linkname, $db_new_linkhref, $db_new_linkorder, $db_new_linkcustomvar, $db_new_linkfontawesome);
-            header("location: " . URL . "/links/");
+            
+            header("location: " . URL . "/links/?action=updated&name=" . $db_new_linkname . "&type=link");
+            die();
+        }
+
+        // Edit Header Submit
+        if(isset($_POST['eh_submit'])) {
+            $db_new_hname = htmlentities($_POST['HeaderName']);
+            $db_new_hlink = "#";
+            $db_new_horder = htmlentities($_POST['HeaderOrder']);
+            $db_new_hddname = addslashes($_POST['HeaderDDName']);
+            $db_new_hfonta = htmlentities($_POST['HeaderFontAwesome']);
+            $home->UpdateHeader($id, $db_new_hname, $db_new_hlink, $db_new_horder, $db_new_hddname, $db_new_hfonta);
+            header("location: " . URL . "/links/?action=updated&name=" . $db_new_hname . "&type=header");
             die();
         }
 
@@ -128,6 +145,43 @@
                         <input type="text" class="form-control" id="LinkCustomVar" name="LinkCustomVar" placeholder="rel=&quot;nofollow&quot; target=&quot;_blank&quot;" value="<?php print(htmlentities($db_edit_linkcustomvar));?>">
                     </div>
                     <button type="submit" class="btn btn-primary float-right" id="el_submit" name="el_submit">Save</button>
+                </form>
+            </div>
+        </div>
+    <?php } ?>
+
+    <?php /* Edit Header */ if($type == "header" && $action == "edit") { ?>
+        <div class="row">
+            <div class="col">
+            <h4>Edit Header</h4>
+            <hr>
+                <form action="" method="post">
+                    <div class="row">
+                        <div class="col">
+                            <div class="form-group">
+                                <label for="HeaderOrder">Header Order</label>
+                                <select class="form-control" id="HeaderOrder" name="HeaderOrder">
+                                    <?php $headerCount = $home->GetHeaderCount(); $hc = 1; while($hc < $headerCount + 1) { ?>
+                                    <option <?php if($db_edit_horder == $hc) { print("selected"); }?>><?php print($hc);?></option>
+                                    <?php $hc++; } ?>
+                                </select>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="form-group">
+                        <label for="HeaderName">Header Name</label>
+                        <input type="text" class="form-control" id="HeaderName" name="HeaderName" placeholder="MyHomeBoardPHP" value="<?php print($db_edit_hname);?>" required>
+                    </div>
+                    <div class="form-group">
+                        <label for="HeaderFontAwesome">Header Font Awesome</label>
+                        <input type="text" class="form-control" id="HeaderFontAwesome" name="HeaderFontAwesome" placeholder="fab fa-github" value="<?php print($db_edit_hfonta);?>">
+                    </div>
+                    <div class="form-group">
+                        <label for="HeaderDDName">Header Drop Down Name</label>
+                        <input type="text" class="form-control" id="HeaderDDName" name="HeaderDDName" placeholder="dropdownDevelopment" value="<?php print($db_edit_hddname);?>" required>
+                        <label for="HeaderDDName" class="text-muted mt-1" style="font-size: 12px;">This is required, suggestion: dropdownHeaderName</label>
+                    </div>
+                    <button type="submit" class="btn btn-primary float-right" id="eh_submit" name="eh_submit">Save</button>
                 </form>
             </div>
         </div>